[mlir][nvvm] move OptimizeForNVVM to its own directory
Cf. https://discourse.llvm.org/t/rfc-separate-gpu-specific-transforms-from-llvmirtransforms/91151
