+++ mlir/include/mlir/Dialect/NVVM/Transforms/OptimizeForNVVM.h
@@ -1,4 +1,4 @@
-//===- OptimizeForNVVM.h - Optimize LLVM IR for NVVM -*- C++ -*-===//
+//===- OptimizeForNVVM.h - Optimize LLVM IR for NVVM ------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,17 +6,17 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef MLIR_DIALECT_LLVMIR_TRANSFORMS_OPTIMIZENVVM_H
-#define MLIR_DIALECT_LLVMIR_TRANSFORMS_OPTIMIZENVVM_H
+#ifndef MLIR_DIALECT_NVVM_TRANSFORMS_OPTIMIZEFORNVVM_H
+#define MLIR_DIALECT_NVVM_TRANSFORMS_OPTIMIZEFORNVVM_H
 
 #include <memory>
 
 namespace mlir {
 class Pass;
-namespace LLVM {
+namespace NVVM {
 #define GEN_PASS_DECL_NVVMOPTIMIZEFORTARGETPASS
-#include "mlir/Dialect/LLVMIR/Transforms/Passes.h.inc"
-} // namespace LLVM
+#include "mlir/Dialect/NVVM/Transforms/Passes.h.inc"
+} // namespace NVVM
 } // namespace mlir
 
-#endif // MLIR_DIALECT_LLVMIR_TRANSFORMS_OPTIMIZENVVM_H
+#endif // MLIR_DIALECT_NVVM_TRANSFORMS_OPTIMIZEFORNVVM_H
