+++ mlir/lib/Dialect/NVVM/Transforms/CMakeLists.txt
@@ -1,0 +1,16 @@
+add_mlir_dialect_library(MLIRNVVMTransforms
+  OptimizeForNVVM.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/NVVM
+
+  DEPENDS
+  MLIRNVVMPassIncGen
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+  MLIRLLVMDialect
+  MLIRNVVMDialect
+  MLIRPass
+  MLIRTransforms
+  )
